pyproject.toml
[tool.poetry]
name = "markdown-search"
version = "0.1.0"
description = ""
authors = ["Andrey Vasnetsov <andrey@vasnetsov.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
mistletoe = "^1.0.1"
openai = "^0.27.2"
beautifulsoup4 = "^4.11.2"
qdrant-client = "^1.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
